more renames  [INTERNAL_BRANCH=sean/webserver-dagit-app-rename]
