+++ .buildkite/dagster-buildkite/dagster_buildkite/steps/dagster_ui.py
@@ -29,7 +29,7 @@ def build_dagster_ui_steps() -> List[CommandStep]:
     return [
         CommandStepBuilder(":typescript: dagster-ui")
         .run(
-            "cd js_modules/dagit",
+            "cd js_modules/dagster-ui",
             # Explicitly install Node 16.x because BK is otherwise running 12.x.
             # Todo: Fix BK images to use newer Node versions, remove this.
             "curl -sL https://deb.nodesource.com/setup_16.x | bash -",
