+++ docs/content/community/contributing.mdx
@@ -82,11 +82,11 @@ dagster-webserver -p 3333 -f complex_pipeline.py
 Keep this running. Then, in another terminal, run the local development (autoreloading, etc.) version of the webapp:
 
 ```bash
-cd dagster/js_modules/dagit
+cd dagster/js_modules/dagster-ui
 make dev_webapp
 ```
 
-During development, you might find these commands useful. Run them from `dagster/js_modules/dagit`:
+During development, you might find these commands useful. Run them from `dagster/js_modules/dagster-ui`:
 
 - `yarn ts`: Typescript typechecking
 - `yarn lint`: Linting with autofix
