+++ docs/next/package.json
@@ -69,7 +69,7 @@
     "@types/node": "^14.14.22",
     "@types/react": "^17.0.0",
     "eslint": "^8.45.0",
-    "eslint-plugin-dagster-rules": "link:../../js_modules/dagit/packages/eslint-config/rules",
+    "eslint-plugin-dagster-rules": "link:../../js_modules/dagster-ui/packages/eslint-config/rules",
     "html-react-parser": "^2.0.0",
     "jest": "^29.6.2",
     "node-html-parser": "^5.3.3",
