+++ js_modules/dagster-ui/package.json
@@ -3,12 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "build": "yarn workspace @dagster-io/dagster-app-oss build && yarn post-build",
+    "build": "yarn workspace @dagster-io/app-oss build && yarn post-build",
     "build-with-profiling": "yarn workspace @dagster-io/app-oss build --profile && yarn post-build",
     "post-build": "cd ../../python_modules/dagster-webserver/dagster_webserver && rm -rf webapp && mkdir -p webapp && cp -r ../../../js_modules/dagster-ui/packages/app-oss/build ./webapp/ && mkdir -p webapp/build/vendor && cp -r graphql-playground ./webapp/build/vendor",
     "lint": "yarn workspace @dagster-io/app-oss lint && yarn workspace @dagster-io/ui-core lint && yarn workspace @dagster-io/ui-components lint",
     "start": "yarn workspace @dagster-io/app-oss start",
-    "ts": "yarn workspace @dagster-io/app-oss && yarn workspace @dagster-io/ui-components ts"
+    "ts": "yarn workspace @dagster-io/app-oss ts && yarn workspace @dagster-io/ui-components ts"
   },
   "workspaces": {
     "packages": [
