+++ js_modules/dagster-ui/packages/app-oss/package.json
@@ -9,8 +9,8 @@
     "@blueprintjs/core": "^4.17.8",
     "@blueprintjs/popover2": "1.13.12",
     "@blueprintjs/select": "^4.9.12",
-    "@dagster-io/ui-core": "workspace:*",
     "@dagster-io/ui-components": "workspace:*",
+    "@dagster-io/ui-core": "workspace:*",
     "@rive-app/react-canvas": "^3.0.34",
     "@types/styled-components": "^5.1.26",
     "graphql": "^16.6.0",
