+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetEvents.tsx
@@ -1,4 +1,11 @@
-import {Box, ButtonGroup, Colors, Spinner, Subheading, ErrorBoundary} from '@dagster-io/ui-components';
+import {
+  Box,
+  ButtonGroup,
+  Colors,
+  Spinner,
+  Subheading,
+  ErrorBoundary,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 
 import {LiveDataForNode, stepKeyForAsset} from '../asset-graph/Utils';
