+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetNodeLineage.tsx
@@ -1,4 +1,12 @@
-import {Box, Button, ButtonGroup, Colors, Icon, JoinedButtons, TextInput} from '@dagster-io/ui-components';
+import {
+  Box,
+  Button,
+  ButtonGroup,
+  Colors,
+  Icon,
+  JoinedButtons,
+  TextInput,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 import styled from 'styled-components';
