+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetPageHeader.tsx
@@ -1,6 +1,14 @@
 // eslint-disable-next-line no-restricted-imports
 import {BreadcrumbProps, Breadcrumbs} from '@blueprintjs/core';
-import {Box, Colors, PageHeader, Heading, Icon, Tooltip, IconWrapper} from '@dagster-io/ui-components';
+import {
+  Box,
+  Colors,
+  PageHeader,
+  Heading,
+  Icon,
+  Tooltip,
+  IconWrapper,
+} from '@dagster-io/ui-components';
 import React from 'react';
 import {Link} from 'react-router-dom';
 import styled from 'styled-components';
