+++ js_modules/dagster-ui/packages/ui-core/src/assets/AutoMaterializePolicyPage/AssetKeysDialog.tsx
@@ -1,4 +1,12 @@
-import {Box, Button, Colors, Dialog, DialogFooter, NonIdealState, TextInput} from '@dagster-io/ui-components';
+import {
+  Box,
+  Button,
+  Colors,
+  Dialog,
+  DialogFooter,
+  NonIdealState,
+  TextInput,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 
 interface Props {
