+++ js_modules/dagster-ui/packages/ui-core/src/assets/CurrentRunsBanner.tsx
@@ -1,5 +1,5 @@
 import {Alert, Box, Spinner} from '@dagster-io/ui-components';
-import {BorderSetting} from '@dagster-io/ui/src/components/types';
+import {BorderSetting} from '@dagster-io/ui-components/src/components/types';
 import React from 'react';
 import {Link} from 'react-router-dom';
