+++ js_modules/dagster-ui/packages/ui-core/src/assets/FailedRunSinceMaterializationBanner.tsx
@@ -1,5 +1,5 @@
 import {Alert, Box} from '@dagster-io/ui-components';
-import {BorderSetting, DirectionalSpacing} from '@dagster-io/ui/src/components/types';
+import {BorderSetting, DirectionalSpacing} from '@dagster-io/ui-components/src/components/types';
 import React from 'react';
 import {Link} from 'react-router-dom';
