+++ js_modules/dagster-ui/packages/ui-core/src/launchpad/LaunchRootExecutionButton.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import {IconName} from '../../../ui/src';
+import {IconName} from '../../../ui-components/src';
 import {LaunchBehavior} from '../runs/RunUtils';
 import {LaunchPipelineExecutionMutationVariables} from '../runs/types/RunUtils.types';
