+++ js_modules/dagster-ui/packages/ui-core/src/launchpad/LaunchpadSessionError.tsx
@@ -1,4 +1,9 @@
-import {NonIdealState, NonIdealStateProps, Spinner, SplitPanelContainer} from '@dagster-io/ui-components';
+import {
+  NonIdealState,
+  NonIdealStateProps,
+  Spinner,
+  SplitPanelContainer,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 
 import {SessionSettingsBar} from './SessionSettingsBar';
