+++ js_modules/dagster-ui/packages/ui-core/src/metadata/TableSchema.tsx
@@ -1,6 +1,6 @@
 import {gql} from '@apollo/client';
 import {Box, Colors, Tag, Tooltip} from '@dagster-io/ui-components';
-import {Spacing} from '@dagster-io/ui/src/components/types';
+import {Spacing} from '@dagster-io/ui-components/src/components/types';
 import * as React from 'react';
 import styled from 'styled-components';
