+++ js_modules/dagster-ui/packages/ui-core/src/nav/JobMetadata.tsx
@@ -1,5 +1,13 @@
 import {gql, useQuery} from '@apollo/client';
-import {Box, Button, ButtonLink, Colors, DialogFooter, Dialog, Tag} from '@dagster-io/ui-components';
+import {
+  Box,
+  Button,
+  ButtonLink,
+  Colors,
+  DialogFooter,
+  Dialog,
+  Tag,
+} from '@dagster-io/ui-components';
 import uniq from 'lodash/uniq';
 import * as React from 'react';
 import {Link} from 'react-router-dom';
