+++ js_modules/dagster-ui/packages/ui-core/src/nav/RepositoryLink.tsx
@@ -1,4 +1,12 @@
-import {Box, Colors, Icon, IconWrapper, MiddleTruncate, Spinner, Tooltip} from '@dagster-io/ui-components';
+import {
+  Box,
+  Colors,
+  Icon,
+  IconWrapper,
+  MiddleTruncate,
+  Spinner,
+  Tooltip,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 import {Link} from 'react-router-dom';
 import styled from 'styled-components';
