+++ js_modules/dagster-ui/packages/ui-core/src/nav/ScheduleOrSensorTag.tsx
@@ -1,4 +1,12 @@
-import {Box, ButtonLink, Colors, Tag, Tooltip, FontFamily, MiddleTruncate} from '@dagster-io/ui-components';
+import {
+  Box,
+  ButtonLink,
+  Colors,
+  Tag,
+  Tooltip,
+  FontFamily,
+  MiddleTruncate,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 import {Link} from 'react-router-dom';
