+++ js_modules/dagster-ui/packages/ui-core/src/partitions/JobBackfillsTable.tsx
@@ -1,5 +1,10 @@
 import {gql, useQuery} from '@apollo/client';
-import {Box, CursorPaginationControls, CursorPaginationProps, NonIdealState} from '@dagster-io/ui-components';
+import {
+  Box,
+  CursorPaginationControls,
+  CursorPaginationProps,
+  NonIdealState,
+} from '@dagster-io/ui-components';
 import React from 'react';
 
 import {RepositorySelector} from '../graphql/types';
