+++ js_modules/dagster-ui/packages/ui-core/src/pipelines/GraphExplorer.tsx
@@ -1,7 +1,13 @@
 import {gql} from '@apollo/client';
 // eslint-disable-next-line no-restricted-imports
 import {Breadcrumbs} from '@blueprintjs/core';
-import {Checkbox, Colors, SplitPanelContainer, TextInput, ErrorBoundary} from '@dagster-io/ui-components';
+import {
+  Checkbox,
+  Colors,
+  SplitPanelContainer,
+  TextInput,
+  ErrorBoundary,
+} from '@dagster-io/ui-components';
 import Color from 'color';
 import qs from 'qs';
 import * as React from 'react';
