+++ js_modules/dagster-ui/packages/ui-core/src/plugins/sql.tsx
@@ -1,4 +1,10 @@
-import {Button, DialogFooter, Dialog, Icon, StyledReadOnlyCodeMirror} from '@dagster-io/ui-components';
+import {
+  Button,
+  DialogFooter,
+  Dialog,
+  Icon,
+  StyledReadOnlyCodeMirror,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 
 import {IPluginSidebarProps} from '../plugins';
