+++ js_modules/dagster-ui/packages/ui-core/src/runs/LogFilterSelect.tsx
@@ -1,4 +1,13 @@
-import {Box, Button, Checkbox, Colors, Icon, Menu, MenuItem, Popover} from '@dagster-io/ui-components';
+import {
+  Box,
+  Button,
+  Checkbox,
+  Colors,
+  Icon,
+  Menu,
+  MenuItem,
+  Popover,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 import styled from 'styled-components';
