+++ js_modules/dagster-ui/packages/ui-core/src/sensors/SensorStateChangeDialog.tsx
@@ -1,7 +1,15 @@
 import {useMutation} from '@apollo/client';
 // eslint-disable-next-line no-restricted-imports
 import {ProgressBar} from '@blueprintjs/core';
-import {Button, Colors, DialogBody, DialogFooter, Dialog, Group, Icon} from '@dagster-io/ui-components';
+import {
+  Button,
+  Colors,
+  DialogBody,
+  DialogFooter,
+  Dialog,
+  Group,
+  Icon,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 
 import {
