+++ js_modules/dagster-ui/packages/ui-core/src/ui/NotebookButton.tsx
@@ -1,4 +1,11 @@
-import {Button, DialogBody, DialogFooter, Dialog, Icon, ExternalAnchorButton} from '@dagster-io/ui-components';
+import {
+  Button,
+  DialogBody,
+  DialogFooter,
+  Dialog,
+  Icon,
+  ExternalAnchorButton,
+} from '@dagster-io/ui-components';
 import * as React from 'react';
 
 import {AppContext} from '../app/AppContext';
