+++ js_modules/dagster-ui/packages/ui-core/src/ui/SectionedLeftNav.tsx
@@ -1,4 +1,12 @@
-import {BaseTag, Box, Colors, Icon, IconWrapper, MiddleTruncate, StyledTag} from '@dagster-io/ui-components';
+import {
+  BaseTag,
+  Box,
+  Colors,
+  Icon,
+  IconWrapper,
+  MiddleTruncate,
+  StyledTag,
+} from '@dagster-io/ui-components';
 import {useVirtualizer} from '@tanstack/react-virtual';
 import * as React from 'react';
 import {useRouteMatch} from 'react-router-dom';
