+++ js_modules/dagster-ui/packages/ui-core/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "paths": {
-      "@dagster-io/ui-components": ["../ui/src"]
+      "@dagster-io/ui-components": ["../ui-components/src"]
     },
     "module": "es2022",
     "target": "es2022",
