+++ python_modules/dagster-webserver/build_js.sh
@@ -1,6 +1,6 @@
 #! /bin/bash
 ROOT=$(git rev-parse --show-toplevel)
-pushd "$ROOT/js_modules/dagit"
+pushd "$ROOT/js_modules/dagster-ui"
 set -eux
 
 yarn install
